Fixes runtime errors
